requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
qdrant-client==1.11.1
openai==1.40.6
pymupdf==1.24.9
pdfplumber==0.11.4
pillow==10.4.0
pytesseract==0.3.13
rapidfuzz==3.9.6
